wip new period bot task
